refactor: input 검증 로직을 InputValidator로 옮김
