+++ src/test/java/baseball/util/InputValidatorTest.java
@@ -6,13 +6,13 @@ import org.junit.jupiter.params.provider.ValueSource;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-class DuplicateCheckerTest {
+class InputValidatorTest {
 
     @DisplayName("같은 숫자가 있으면 예외를 발생시킨다.")
     @ParameterizedTest
     @ValueSource(strings = {"111", "999", "122", "991", "121"})
-    void hasSameNumber(String input) {
-        assertThatThrownBy(() -> DuplicateChecker.hasSameNumber(input))
+    void validateSameNumber(String input) {
+        assertThatThrownBy(() -> InputValidator.validate(input))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 }
